[chore] Presentational & Container 패턴 적용
폴더 구조 변경, 파일 수정
