+++ src/components/asset/components/presentational/AssetList.js
@@ -1,32 +1,8 @@
 import React from 'react';
-import { Route, Link, useRouteMatch } from 'react-router-dom';
-import '../../../styles/sass/main.css';
-import stock_icon from '../../../styles/images/stock_icon.png';
-import btc_icon from '../../../styles/images/btc_icon.png';
+import { Link } from 'react-router-dom';
+import '../../../../styles/sass/main.css';
 
-const AssetList = () => {
-  let match = useRouteMatch();
-  const AssetBox = [
-    {
-      title: '주식',
-      link: 'stock',
-      img: stock_icon,
-      total: '190,000 TYL',
-      today: '+',
-      value: '4,580',
-      percent: '5.1',
-    },
-    {
-      title: '비트코인',
-      link: 'btc',
-      img: btc_icon,
-      total: '111,000 TYL',
-      today: '-',
-      value: '10,580',
-      percent: '8.4',
-    },
-  ];
-
+const AssetList = ({ AssetBox, match }) => {
   const AssetList = AssetBox.map((menu, idx) => (
     <li key={idx}>
       <Link to={`${match.url}/${menu.link}`}>
