+++ src/components/profile/components/presentational/Modal.js
@@ -1,5 +1,5 @@
-import React, { useEffect } from 'react';
-import '../../../styles/sass/main.css';
+import React from 'react';
+import '../../../../styles/sass/main.css';
 
 const Modal = props => {
   // 열기, 닫기, 모달 헤더 텍스트를 부모로부터 받아옴
